+++ common/blockchain/broadcaster.go
@@ -40,7 +40,15 @@ func (self Broadcaster) Broadcast(tx *types.Transaction) (map[string]error, bool
 	wg.Wait()
 	result := map[string]error{}
 	failures.Range(func(key, value interface{}) bool {
-		result[key.(string)] = value.(error)
+		k, ok := key.(string)
+		if !ok {
+			return false
+		}
+		err, ok := value.(error)
+		if !ok {
+			return false
+		}
+		result[k] = err
 		return true
 	})
 	return result, len(result) != len(self.clients) && len(self.clients) > 0