+++ exchange/binance_format.go
@@ -2,6 +2,7 @@ package exchange
 
 import (
 	"encoding/json"
+	"fmt"
 )
 
 type Binaprice struct {
@@ -12,8 +13,16 @@ type Binaprice struct {
 func (self *Binaprice) UnmarshalJSON(text []byte) error {
 	temp := []interface{}{}
 	err := json.Unmarshal(text, &temp)
-	self.Quantity = temp[1].(string)
-	self.Rate = temp[0].(string)
+	qty, ok := temp[1].(string)
+	if !ok {
+		return fmt.Errorf("Unmarshal err: interface %v can't be converted to string", temp[1])
+	}
+	self.Quantity = qty
+	rate, ok := temp[0].(string)
+	if !ok {
+		return fmt.Errorf("Unmarshal err: interface %v can't be converted to string", temp[0])
+	}
+	self.Rate = rate
 	return err
 }