+++ stat/reserve_stat.go
@@ -218,7 +218,11 @@ func (self ReserveStats) GetHeatMap(fromTime, toTime uint64, tzparam int64) (com
 			return arrResult, err
 		}
 		for _, stat := range cStats {
-			s := stat.(common.MetricStats)
+			s, ok := stat.(common.MetricStats)
+			if !ok {
+				err = fmt.Errorf("Can not convert stat (%v) to MetricStat", s)
+				continue
+			}
 			current := result[c]
 			result[c] = common.HeatmapType{
 				TotalETHValue:        current.TotalETHValue + s.ETHVolume,
@@ -269,7 +273,11 @@ func (self ReserveStats) GetTokenHeatmap(fromTime, toTime uint64, tokenStr, freq
 			return arrResult, err
 		}
 		for _, stat := range stats {
-			s := stat.(common.VolumeStats)
+			s, ok := stat.(common.VolumeStats)
+			if !ok {
+				err = fmt.Errorf("Can not convert stat (%v) to VolumeStats", s)
+				continue
+			}
 			current := result[country]
 			result[country] = common.VolumeStats{
 				Volume:    current.Volume + s.Volume,
@@ -506,7 +514,11 @@ func (self ReserveStats) ExceedDailyLimit(address ethereum.Address) (bool, error
 				log.Printf("Got more than 1 day stats. This is a bug in GetUserVolume")
 			} else {
 				for _, volume := range volumeStats {
-					volumeValue := volume.(common.VolumeStats)
+					volumeValue, ok := volume.(common.VolumeStats)
+					if !ok {
+						log.Printf("Can not convert volume (%v) to VolumeStats", volume)
+						continue
+					}
 					totalVolume += volumeValue.USDAmount
 					break
 				}