+++ Changelog.md
@@ -9,6 +9,7 @@
 - Enable errcheck checker and fix all unhandled errors (#307)
 - Fix GetAccounts error is ignored for houbi exchange (#308)
 - Fix missing required fields in PWIs v2 APIs (#314)
+- Fix type/interface conversion error swallow. Now these errors will be printed out.
 
 ### Improvements: